clean up scoring script (#24)
* clean up scoring script

* Apply suggestions from code review


---------
