+++ run.sh
@@ -16,14 +16,19 @@ chmod +x run.sh
 ./run.sh
 END_COMMENT
 
-## create folder structure
-export data_split="test"
-export task_folder="test_bioclip"
-export ref_path="/home/wu.5686/imageo/challenge/reference_data/ref_$data_split.csv"
 
-mkdir -p sample_result_submission/$task_folder/ref
-mkdir -p sample_result_submission/$task_folder/res
-cp $ref_path sample_result_submission/$task_folder/ref
+export task_type="evaluate" #folder; predict; evaluate
+export data_split="dev"
+export task_folder="dev_dino2"
+
+
+## create folder structure
+if [[ "$task_type" == *"folder"* ]]; then
+    mkdir -p sample_result_submission/$task_folder/ref
+    mkdir -p sample_result_submission/$task_folder/res
+    export ref_path="/home/wu.5686/imageo/challenge/reference_data/ref_$data_split.csv"
+    cp $ref_path sample_result_submission/$task_folder/ref
+fi
 
 : <<'END_COMMENT'
 now, the folder structure for task is as below:
@@ -35,20 +40,19 @@ sample_result_submission
 END_COMMENT
 
 ## get the predictions
-export input_dir="input_data/$data_split"
-# export input_dir="/local/scratch/wu.5686/anomaly_challenge/input_data/dev_A"
-export output_dir="sample_result_submission/$task_folder/res"
-export program_dir="ingestion_program"
-# export submission_dir="baselines/DINO_SGD_code_submission"
-export submission_dir="baselines/BioCLIP_code_submission"
-python ingestion_program/ingestion.py $input_dir $output_dir $program_dir $submission_dir
+if [[ "$task_type" == *"predict"* ]]; then
+    export input_dir="input_data/$data_split"
+    # export input_dir="/local/scratch/wu.5686/anomaly_challenge/input_data/$data_split"
+    export output_dir="sample_result_submission/$task_folder/res"
+    export program_dir="ingestion_program"
+    export submission_dir="baselines/DINO_SGD_code_submission"
+    # export submission_dir="baselines/BioCLIP_code_submission"
+    python ingestion_program/ingestion.py $input_dir $output_dir $program_dir $submission_dir
+fi
 
 ## score the predictions
-export input_dir="sample_result_submission/$task_folder"
-export output_dir="sample_result_submission/$task_folder"
-python scoring_program/score_combined.py $input_dir $output_dir
-
-
-
-
-
+if [[ "$task_type" == *"evaluate"* ]]; then
+    export input_dir="sample_result_submission/$task_folder"
+    export output_dir="sample_result_submission/$task_folder"
+    python scoring_program/score_combined.py $input_dir $output_dir
+fi
